refactor(repository/, entity/, dto/, LocationService): change id's type for entity from Integer to Long
* adapt repositories
* adapt dtos
* adapt LocationService

+++ src/main/java/com/ddevuss/weather/oracle/dto/WeatherForecastDto.java
@@ -9,7 +9,7 @@ import java.io.Serializable;
 @Builder
 public class WeatherForecastDto implements Serializable {
 
-    Integer locationId;
+    Long locationId;
     String locationName;
     String countryCode;
     String description;
